refactor: View를 LottoInputView, LottoOutputView로 분리
