+++ src/main/java/controller/LottoController.java
@@ -5,20 +5,21 @@ import domain.LottoNumber;
 import domain.Lottos;
 import domain.WinLottoNumbers;
 import util.Parser;
-import view.View;
+import view.lotto.LottoInputView;
+import view.lotto.LottoOutputView;
 
 import java.util.List;
 
 public class LottoController {
 
     public static void run() {
-        Lottos lottos = LottoFactory.createLottos(View.inputAmount());
-        View.printLottos(lottos);
+        Lottos lottos = LottoFactory.createLottos(LottoInputView.inputAmount());
+        LottoOutputView.printLottos(lottos);
 
-        List<LottoNumber> lottoNumbers = Parser.parsingWinNumbers(View.inputWinNumbers());
-        LottoNumber bonusLottoNumber = LottoNumber.getLottoNumber(View.inputBounsNumber());
+        List<LottoNumber> lottoNumbers = Parser.parsingWinNumbers(LottoInputView.inputWinNumbers());
+        LottoNumber bonusLottoNumber = LottoNumber.getLottoNumber(LottoInputView.inputBounsNumber());
         WinLottoNumbers winLottoNumbers = new WinLottoNumbers(lottoNumbers, bonusLottoNumber);
 
-        View.printResult(lottos, winLottoNumbers);
+        LottoOutputView.printResult(lottos, winLottoNumbers);
     }
 }
